Removed unused imports, split LoginController, added cancel button to new loan app.
