+++ src/main/java/com/example/ist412group4/controller/CustomerController.java
@@ -1,10 +1,8 @@
 package com.example.ist412group4.controller;
 
 
-import com.example.ist412group4.model.Employee;
 import com.example.ist412group4.model.Customer;
 import com.example.ist412group4.service.CustomerService;
-import com.example.ist412group4.service.EmployeeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -13,47 +11,10 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 
 @Controller
-public class LoginController {
+public class CustomerController {
 
     @Autowired
     CustomerService customerService;
-
-    @Autowired
-    EmployeeService employeeService;
-
-    @GetMapping("/")
-    public String viewHomePage(Model model) {
-        //model.addAttribute("listCourses", courseService.getAllCourses());
-        return "index";
-    }
-
-    @GetMapping("/showEmployeeLogin")
-    public String viewEmployeeLogin(Model model) {
-        Employee employee = new Employee();
-        model.addAttribute("employee", employee);
-        return "employee_login";
-    }
-
-    @GetMapping("/showNewEmployeeForm")
-    public String showNewEmployeeForm(Model model) {
-        Employee employee = new Employee();
-        model.addAttribute("employee", employee);
-        return "new_employee";
-    }
-
-    @GetMapping("/validateEmployeeLogin")
-    public String validateEmployeeLogin(@ModelAttribute("employee") Employee employee) {
-        if (employeeService.validate(employee)==true){
-            return "employee_menu";
-        } else {
-            return "employee_login_error";
-        }
-    }
-
-    @GetMapping("/employeeMenu")
-    public String viewEmployeeMenu(Model model){
-       return "employee_menu";
-    }
 
     @GetMapping("/showCustomerLogin")
     public String viewCustomerLogin(Model model) {
@@ -90,13 +51,6 @@ public class LoginController {
         return "new_customer_error";
     }
 
-    @PostMapping("/saveEmployee")
-    public String saveEmployee(@ModelAttribute("employee") Employee employee){
-        if (employee.authenticate() == true) {
-            employeeService.saveEmployee(employee);
-            return "redirect:/showEmployeeLogin";
-        }
-        return "new_employee_error";
-    }
-
 }
+
+
