+++ src/main/java/com/example/ist412group4/service/CustomerService.java
@@ -1,7 +1,6 @@
 package com.example.ist412group4.service;
 
 import com.example.ist412group4.model.Customer;
-import org.springframework.data.domain.Page;
 
 import java.util.List;
