+++ src/main/java/com/example/ist412group4/service/CustomerServiceImpl.java
@@ -3,7 +3,6 @@ package com.example.ist412group4.service;
 import java.util.List;
 import java.util.Optional;
 import com.example.ist412group4.model.Customer;
-import com.example.ist412group4.model.Employee;
 import com.example.ist412group4.repository.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
