+++ src/main/resources/application.properties
@@ -1,7 +1,7 @@
 # DATASOURCE (DataSourceAutoConfiguration & DataSourceProperties)
-spring.datasource.url=jdbc:mysql://localhost:3306/eFinance?useSSL=false&serverTimezone=UTC&useLegacyDatetimeCode=false&allowPublicKeyRetrieval=true
-spring.datasource.username=root
-spring.datasource.password=Captain13!
+spring.datasource.url=jdbc:mysql://localhost:3306/efinance?useSSL=false&serverTimezone=UTC&useLegacyDatetimeCode=false&allowPublicKeyRetrieval=true
+spring.datasource.username=nitsrik
+spring.datasource.password=GunnerFuego
 # Hibernate
 # The SQL dialect makes Hibernate generate better SQL for the chosen database
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5InnoDBDialect
