+++ src/main/resources/templates/new_loan_application.html
@@ -48,5 +48,11 @@
         </div>
     </form>
 </div>
+</div>
+<div class="container my-2">
+    <br/>
+    <a th:href="@{/customerMenu}" class="btn btn-primary"> Cancel </a>
+    <br/>
+</div>
 </body>
 </html>
